requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
gunicorn>=21.2.0
python-decouple>=3.8
whitenoise>=6.6.0
django-countries>=7.6.1
drf_spectacular>=0.28.0